+++ skiaChart.d.ts
@@ -1,2 +1,2 @@
-export * from './lib/typescript/skiaChart';
-export { default } from './lib/typescript/skiaChart';
+export * from './lib/typescript/skia/skiaChart';
+export { default } from './lib/typescript/skia/skiaChart';
