+++ src/skia/skiaChart.web.tsx
@@ -1,0 +1,2 @@
+export { default } from '../web/chart';
+export { SVGRenderer } from '../svg/SVGRenderer';
