+++ src/svg/SVGCore.ts
