+++ src/svg/SVGRenderer.ts
