+++ src/svg/svgChart.web.tsx
@@ -1,2 +1,2 @@
-export { default } from './web/chart';
+export { default } from '../web/chart';
 export { SVGRenderer } from './SVGRenderer';
