--- src/svgChart.web.tsx
@@ -1,2 +1,0 @@
-export { default } from './web/chart';
-export { SVGRenderer } from './SVGRenderer';
